app/requirements.txt
black
fastapi
pre-commit
pytest
uvicorn
psycopg2-binary
